+++ backend/src/routes/invoiceRoutes.ts
@@ -1,0 +1,14 @@
+import { Router, Request, Response } from 'express';
+import { createInvoice } from '../controllers/paymentController';
+
+const router = Router();
+
+router.post('/create-invoice', async (req: Request, res: Response) => {
+  try {
+    await createInvoice(req, res);
+  } catch (error: any) {
+    res.status(500).json({ error: error.message });
+  }
+});
+
+export default router;
