+++ backend/src/service/xenditClient.ts
@@ -1,12 +1,13 @@
+import { Xendit } from 'xendit-node';
 import dotenv from 'dotenv';
 
 dotenv.config();
 
 const secretApiKey = process.env.XENDIT_SECRET_KEY!;
-const base64EncodedKey = Buffer.from(`${secretApiKey}:`).toString('base64');
 
+const xenditClient = new Xendit({
+  secretKey: secretApiKey,
+  //xenditURL: 'https://api.xendit.co' // Custom base URL
+});
 
-export const xenditHeaders = {
-  'Authorization': `Basic ${base64EncodedKey}`,
-  'Content-Type': 'application/json'
-};
+export const { Invoice } = xenditClient;
